chore: refactor rspack_plugin_devtool
